+++ Assets/Scripts/Bullet/Iris/Iris_Bullet3R.cs
@@ -45,12 +45,12 @@ public class Iris_Bullet3R : Bullet {
 
             if (collision.tag == "Player" + oNum)
             {
-                PlayerManager.instance.Local.CurrentHp -= damage;
+                PlayerManager.instance.Local.CurrentHp -= (short)damage;
 
             }
             if (collision.gameObject.name == "Graze" && collision.transform.parent.tag == "Player" + oNum)
             {
-                PlayerManager.instance.Local.CurrentSkillGage += 1f;
+                PlayerManager.instance.Local.CurrentSkillGage += (short)1f;
             }
         }
     }
