+++ Assets/Scripts/Bullet/Iris/Iris_Bullet4.cs
@@ -54,11 +54,11 @@ public class Iris_Bullet4 : Bullet {
             if (collision.tag == "Player" + oNum)
             {
                 Debug.Log(PlayerManager.instance.Local.CurrentHp);
-                PlayerManager.instance.Local.CurrentHp -= damage;
+                PlayerManager.instance.Local.CurrentHp -= (short)damage;
             }
             if (collision.gameObject.name == "Graze" && collision.transform.parent.tag == "Player" + oNum)
             {
-                PlayerManager.instance.Local.CurrentSkillGage += 1f;
+                PlayerManager.instance.Local.CurrentSkillGage += 1;
             }
         }
     }
